% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_knn_cv.R
\name{my_knn_cv}
\alias{my_knn_cv}
\title{My_knn_cv}
\usage{
my_knn_cv(train, cl, k_nn, k_cv)
}
\arguments{
\item{train}{Is a matrix with no NAs or missing values that is used to train the model.}

\item{cl}{Is the true classification of the training data.}

\item{k_nn}{Is the number of nearest neighbors to include in the cross validation calculation.}

\item{k_cv}{Is the number of folds to use for the cross validation (common Ks are 2,5, and 10).}
}
\value{
Type list with a \code{cv_err} object and the predicted classification \code{class} output.
}
\description{
This function calculates a k-fold cross validation for the k nearest neighbors algorithm.
}
\examples{
set.seed(1)

rand_data <- data.frame(x1 = rnorm(100,0,1),x2 = rnorm(100,2,1))
rand_data_cl <- data.frame(y = rbinom(100,1,.3))
my_knn_cv(rand_data,rand_data_cl,k_nn = 5,k_cv = 5)
my_knn_cv(rand_data,rand_data_cl,k_nn = 5,k_cv = 10)

}
\keyword{prediction}
