% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_lm.R
\name{my_lm}
\alias{my_lm}
\title{My Linear Model (lm) function}
\usage{
my_lm(formula, data)
}
\arguments{
\item{formula}{A formula object \code{Y~X}.}

\item{data}{A data frame or matrix.}
}
\value{
Returns a list object with a model table containing the \code{coefficients},
\code{t-values}, degrees of freedom, and p-values computed from the t-distribution.
These regression coefficients should be exactly the same as the base R lm().
}
\description{
This function fits and summarizes the results of a linear model
estimated with Ordinary Least Squares.
}
\examples{
my_lm(lifeExp ~ gdpPercap,data = gapminder :: gapminder)
my_lm(lifeExp ~ gdpPercap + country ,data = gapminder :: gapminder)

}
\keyword{inference,prediction}
