% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_rf_cv.R
\name{my_rf_cv}
\alias{my_rf_cv}
\title{my_rf_cv}
\usage{
my_rf_cv(k)
}
\arguments{
\item{k}{Is the number of K folds to run.}
}
\value{
A numeric calculation of the \code{cv_err}.
}
\description{
This function calculate a cross validation error for a random forest using the "palmerpenguins" data.
}
\examples{
my_rf_cv(k = 3)
my_rf_cv(k = 5)

}
\keyword{prediction}
