% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_t.test.R
\name{my_t.test}
\alias{my_t.test}
\title{My_t_test}
\usage{
my_t.test(x, alternative, mu)
}
\arguments{
\item{x}{Numeric input data to use in your function.}

\item{alternative}{Character specifying what the alternative hypothesis is: less, greater,two.sided.}

\item{mu}{The null hypothesis or your test.}
}
\value{
Returns a list with the t.test output and a model table and code
\code{Hypothesis, T-statistic, DF, P-value}. It also returns a call of the hypothesis tested.
}
\description{
This function runs a left, right, and two-sided t test.
}
\examples{
my_t.test(gapminder$lifeExp,alternative = "two.sided",mu = 60)
my_t.test(gapminder$lifeExp,alternative = "less",mu = 60)

}
\keyword{inference.}
